[add] Working while loops and simple game logic

+++ ethan_game_logic/LCD.h
@@ -1,8 +1,10 @@
 #ifndef __STM32L476G_DISCOVERY_LCD_H
 #define __STM32L476G_DISCOVERY_LCD_H
+#include "stm32l476xx.h"
 
 #include <stdint.h>
 #include <string.h>
+#include <stdio.h>
 #define bool _Bool
 
 void LCD_Initialization(void);
@@ -12,7 +14,7 @@ void LCD_PIN_Init(void);
 void LCD_Configure(void);
 void LCD_Clear(void);
 
-void LCD_DisplayNum(uint8_t number);
+void LCD_DisplayNum(uint32_t number);
 void LCD_WriteChar(uint8_t* ch, bool point, bool colon, uint8_t position);
 static void LCD_Conv_Char_Seg(uint8_t* c,bool point,bool colon, uint8_t* digit);
 void LCD_Display_Name(void);
